Dialogs: move out of portals so that they can handle text input.
This make them render a bit less nicely on small screens, and may cause
other issues depending on where the dialog is put, but it's better than
the current situation of the bad rendering.

Essentially the issue (I think) is that Portal makes it skip a render
loop which in turn confuses the RN consolidation code.

The better, long term solution is probably put them back in portals and
just don't use input dialogs in them.

Fixes #25.