Fix editing the bottom of notes on iOS
This has been such a game of whack-a-mole, with fixing it on Android
breaking it on the web, and fixing on that breaks on iOS and etc.
We now only use it on iOS, as it should already just be fine on Android
and the web where the platform takes care of it.

This should now be fixed.

Fixes #46